Create a new object called Thing. Ball, Arrow, Consumables all inherit Thing. I also removed Ball.py, Arrow.py, and Consumables.py and put them in a new file called game_objects.py.
